refactor(domain): split container.ts + adopt feature-id helpers + logger migration
