+++ packages/core/src/application/use-cases/applications/resume-application-workflow.use-case.ts
@@ -17,6 +17,7 @@ import type { IWorkflowStepRepository } from '../../ports/output/repositories/wo
 import type { IInteractiveSessionService } from '../../ports/output/services/interactive-session-service.interface.js';
 import type { IInteractiveSessionRepository } from '../../ports/output/repositories/interactive-session-repository.interface.js';
 import type { SendInteractiveMessageUseCase } from '../interactive/send-interactive-message.use-case.js';
+import { featureIdForApplication } from '../../../domain/shared/feature-id.js';
 import { APPLICATION_CREATION_WORKFLOW } from './application-creation.workflow.js';
 
 export interface ResumeApplicationWorkflowInput {
@@ -42,7 +43,7 @@ export class ResumeApplicationWorkflowUseCase {
     const app = await this.appRepo.findById(input.applicationId);
     if (!app) throw new Error(`Application ${input.applicationId} not found`);
 
-    const featureId = `app-${app.id}`;
+    const featureId = featureIdForApplication(app.id);
     const steps = await this.stepRepo.listByFeature(featureId);
     if (steps.length === 0) return;
